add partial runs, update to use rio-tiler cog functions

+++ rezoning_api/db/layers.py
@@ -1,4 +1,5 @@
 """layer functions"""
+import os
 import json
 from time import time
 
@@ -26,25 +27,32 @@ def get_layers():
     # return json.loads(s3_get(BUCKET, 'api/layers.json'))
 
 
-def refresh_country_extrema():
+def refresh_country_extrema(partial=False):
     """refresh the country minima and maxima per layer"""
     layers = get_layers()
     for feature in world["features"]:
+        fname = f"temp/{feature['properties']['GID_0']}.json"
         t1 = time()
+        if partial and os.path.exists(fname):
+            print(f"skipping {fname} already exists")
+            continue
         print(f"reading values for {feature['properties']['NAME_0']}")
-        extrema = dict()
-        for dataset in datasets:
-            print(f"read {dataset}")
-            ds = read_dataset(
-                f"s3://{BUCKET}/multiband/{dataset}.tif",
-                layers[dataset],
-                feature["geometry"],
-            )
-            for layer in layers[dataset]:
-                extrema[layer] = dict(
-                    min=float(ds.sel(layer=layer).min()),
-                    max=float(ds.sel(layer=layer).max()),
+        try:
+            extrema = dict()
+            for dataset in datasets:
+                print(f"read {dataset}")
+                ds = read_dataset(
+                    f"s3://{BUCKET}/multiband/{dataset}.tif",
+                    layers[dataset],
+                    feature,
                 )
-        with open(f"temp/{feature['properties']['GID_0']}.json", "w") as out:
-            json.dump(extrema, out)
+                for layer in layers[dataset]:
+                    extrema[layer] = dict(
+                        min=float(ds.sel(layer=layer).min()),
+                        max=float(ds.sel(layer=layer).max()),
+                    )
+            with open(fname, "w") as out:
+                json.dump(extrema, out)
+        except Exception:
+            print("error, skipping")
         print(f"elapsed: {time() - t1} seconds")
